% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grafico_residuos.R
\name{grafico_residuos}
\alias{grafico_residuos}
\title{Plota graficos de diagnostico dos residuos de um modelo}
\usage{
grafico_residuos(modelo)
}
\arguments{
\item{modelo}{Um modelo estimado.}
}
\value{
Quatro gráficos de diagnóstico
}
\description{
Plota graficos de diagnostico dos residuos de um modelo
}
\examples{
grafico_residuos(lm(y ~ x))
}

