% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rcl_estados}
\alias{rcl_estados}
\title{Receita Corrente Líquida para os Estados Brasileiros}
\format{Um data frame com xx linhas e xx variáveis:
\describe{
  \item{ano}{ano de arrecadação}
  \item{mes}{mês da arrecadação}
}}
\source{
\url{https://www.contaspublicas.caixa.gov.br/}
}
\usage{
rcl_estados
}
\description{
Base de dados contendo a receita corrente líquida para os estados brasileiros
disponível no Relatório Resumido de Execução Orçamentária de cada ente
}
\keyword{datasets}

